+++ src/user-list-item.tsx
@@ -3,7 +3,7 @@ import * as React from 'react';
 import Avatar from 'material-ui/Avatar';
 import { ListItem } from 'material-ui/List';
 
-import { Api } from './lib/models/api-call';
+import { Api } from './lib/models/slack-api';
 import { Model, fromObservable } from './lib/model';
 import { SimpleView } from './lib/view';
 import { Store } from './store';
